+++ frontend/src/App.jsx
@@ -1,5 +1,4 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import { useState } from "react";
 import { scrapeRecipe } from "./services/recipes";
 
 import "./App.css";
@@ -10,7 +9,6 @@ import { SingleRecipePage } from "./pages/RecipePage/SingleRecipePage";
 import { CreateRecipePage } from "./pages/RecipePage/CreateRecipePage";
 import { RecipeCollection } from "./pages/RecipeCollection/RecipeCollection";
 import Navbar from "./components/Navbar";
-import FavouriteButton from "./components/RecipePage/FavouriteButton";
 import RecipeScraper from "./components/RecipeScraper";
 import { useState } from "react";
 import { logout } from "./services/authentication";
@@ -65,7 +63,6 @@ const App = () => {
   return (
     <BrowserRouter>
       <Navbar isLoggedIn={isLoggedIn} onLogout={handleLogout} />
-      <FavouriteButton />
       <Routes>
         <Route
           path="/"
