+++ frontend/src/services/recipes.js
@@ -140,3 +140,26 @@ export const getRecipeById = async (recipeId, token) => {
 
   return data;
 };
+
+export const toggleFavourite = async (recipeId, token) => {
+  console.log('Toggling favourite for recipe ID:', recipeId);
+  const requestOptions = {
+    method: "PATCH",
+    headers: {
+      Authorization: `Bearer ${token}`,
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({ recipeId }),
+};
+console.log('Sending request to server...');
+const response = await fetch(`${BACKEND_URL}/recipes/favouritedByOwner/${recipeId}`, requestOptions);
+
+if (response.status !== 200) {
+  throw new Error("Failed to toggle favourite button")
+}
+
+console.log('Toggle favourite successful')
+  const data = await response.json();
+  console.log('Response data:', data);
+  return data;
+};
